updated synchronization
previously, if a node created a new block, that block would have a certain prevHash. During the synchronization process, there's a chance that another block will have to be added to the blockchain instead, before the node's own block. In that case, the prevHash value must be updated and the hash recalculated. This is exactly what I implemented.